requirements.txt
# ML & Data Science
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
stable-baselines3>=1.5.0
gym>=0.21.0

# Trading & Financial
MetaTrader5>=5.0.0
ta>=0.7.0  # Technical Analysis
backtrader>=1.9.76

# Visualization
matplotlib>=3.4.3
plotly>=5.3.1
mplfinance>=0.12.7

# UI
PyQt5>=5.15.4
QtPy>=1.9.0

# Utils
PyYAML>=5.4.1
python-dotenv>=0.19.0
loguru>=0.5.3

# Testing
pytest>=6.2.5
pytest-qt>=3.3.0